backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
neo4j==5.28.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
